refactor: reorganize helper functions into model-dependent and model-independent to break an import cycle
ENT-11544
