+++ CHANGELOG.rst
@@ -17,6 +17,10 @@ Unreleased
 ----------
 * nothing unreleased
 
+[8.1.1] - 2026-06-18
+---------------------
+* refactor: reorganize helper functions into model-dependent and model-independent to break an import cycle
+
 [8.1.0] - 2026-06-17
 ---------------------
 * feat: add courseware-access pipeline steps
