+++ enterprise/__init__.py
@@ -2,4 +2,4 @@
 Your project description goes here.
 """
 
-__version__ = "8.1.0"
+__version__ = "8.1.1"
